backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
google-api-python-client==2.126.0
google-auth==2.29.0
python-dotenv==1.0.1
# better-profanity==0.7.0
# scikit-learn==1.4.1.post1  # Only needed if you use it elsewhere
# joblib==1.4.0